O_DIRECT is a global flag for the engine, which requires file page allocation sizes be at least as large as the required alignment (4KB on Linux).  We've been relying on the application to do the right thing by configuring allocation_size=XXX as part of the WT_SESSION.create call, but that isn't sufficient, as there's no way to set the page allocation size for our internal files (in the Facebook case, WiredTiger.wt).
Add code to silently change any file allocation size to at least as large as
the buffer alignment we're using, plus, if it's not a multiple of the buffer
alignment, increment it to the next largest multiple of that value.  The
former is known to be necessary, the latter is because I think it's likely to
be necessary and I can't prove it's not.  (And I'm tired of screwing around
with this.)

One minor code shuffle, we've been passing the the file allocation size to the
block manager on create and truncate, doing that same thing for open removes
the last place in the block manager that looks up the allocation size, which
is probably the right simplification to make.