More namespace changes: we can now pass kvs_XXX options (including kvs_devices) to every session method that takes a URI.
Move the URI structure configuration into the session.open_cursor support,
all session.create does now is store the master record for the URI on
the device.  Move format configuration from the cursor into the underlying
URI structure.

Rework the record number allocation to avoid races (we lock it before
we allocate a value, and we check that any explicit record creation
doesn't allocate a record past it).

Sort the list of devices so ordering doesn't affect naming.