SERVER-10437 remove chaining.js test
It is testing functionality that has been deprecated by the new sync source updater.
Also, tags.js covers testing chaining with network failures, so there is no further
need for this test.  chaining.js is racy with respect to the way it activates a
failpoint in the socket class, which was causing the test to fail sometimes.