Fix an LSM bug where cursors were not checking all chunks for existence.
To fire, this bug required that a thread race inside __clsm_open_cursors, then
some of the cursors on the oldest parts of the LSM tree were closed.

In the process of fixing this and checking the cursor state at the end of
__clsm_open_cursors, add a flag to track if chunks are empty on disk without a
checkpoint.  Don't repeatedly reopen all cursors in that case.

refs #670, #671