Test the WT_BTREE.modified value before setting it, it's probably a hot cache line.  (This is the only real change here, the rest is semantic sugar.)
Set the WT_BTREE.modified value after we've flushed the insert/update
information (I don't think this is a bug, but it feels cleaner).

Flip the default for page-dirty: by default, mark both the tree and the
page dirty.  There are 3 places we have reason not to mark the tree
dirty, add comments for them; clean up the places we were setting the
WT_BTREE.modified value explicitly instead of calling the right modify
function.  This doesn't change the previous behavior, but it's safer --
I want people to have to search for some way to NOT dirty the tree, in
the rare case that's useful.