Fix a problem with the cache's dirty-byte calculation; a page can be marked clean or dirty while the page's footprint is being updated (for example, the WT_UPDATE "obsolete" discard function can obviously be updating the page's footprint when the page is clean or dirty).
Handle this by saving a copy of the last value used to increment/decrement
the cache's dirty-byte count for this page, and when the page transitions,
use that saved value to reverse the increment/decrement, not the page's
current footprint.

This means the cache's dirty-byte information lags reality because each
change to the page's footprint won't be reflected in the cache until the
page transitions.  To keep that from being too much of a problem, update
the page's cache dirty-btye information as part of reconciliation, even
if the page isn't transitioning from dirty to clean.

If updating during reconciliation is insufficient, we could also do a
similar information update anywhere we know that only a single thread
of control can being doing the update.  For example, when a
serialization function calls __wt_page_modify_set(), we could check and
if the saved footprint value is sufficiently different from the current
footprint value, we could do the same information update there, since
only a single thread can be doing a serialized update on a page at any
time.