__rec_review() goes recursive, don't clear the istree state flag on each call.
I made the same change for inmem_split.  It isn't a bug, because the
check for "can this page be merged" will cause us to return EBUSY before
we get to the code that flags an in-memory split, but it's confusing for
a recursive function to clear the flag on each call.