The stat "clear" function isn't correct for initialization because it ignores some fields, its intent is to clear specific fields between stat calls, not clear the whole structure.  Add clear functionality to the stat initialization function so it can be used to refresh stat structures.
Change the name of the stat "clear" function to "refresh" so nobody
makes this mistake in the future (but don't drive that change out to the
application-level, "clear" is still the right way for the application
programmer to think about this).

Change the LSM and the table statistics aggregation code to always call
the statistics structure's init function, that will clear any previous
values.