A different approach to aggregating checkpoint handle statistics; when retrieving the cursor statistics for a file URI, run through the list of open handles and aggregate statistics for any checkpoint handles on the same underlying URI.
This backs out most of the previous changes to conn/conn_stat.c on this
branch.

Some notable changes along the way:

Don't bother doing next/prev on the statistics cursor in the stat log
server code, we know where the values are, just get them.

When opening the statistics cursor in the stat log server code, use a
default cursor configuration.

This means btree cursors potentially require a copy into a separate
WT_DSRC_STATS structure (just like tables and LSM); bite that bullet and
do it all the time both for btree cursors and connection statistics
cursors.  That means we (1) don't need to pin the underlying btree but
can immediately release it, and (2) don't have to remember how to clear
the underlying statistics when we close the statistics cursor, we can
clear them immediately after the copy.