SERVER-10564 plan cache data storage mechanism
The plan cache now builds a tree isomorphic to the tagged match
expression in order to store the index tags used by a plan. On
cache hit, this tree is used to tag the match expression. The tagged
tree can then be fed through the access planner in order to recreate
the original query solution.