Rework the pthread spinlock logging code.
Split the registry of the spinlock and the registry of a calling
location, they're really completely separate structures, they just
happen to be used for similar purposes

LSM can open a lot of files, increase the number of spinlocks we can
track to 1024, and the maximum spinlock caller location ID's to 60.

Aggregate the block manager and btree page spinlock information, LSM
with bloom filters can open a lot of files.