Minor re-working of 9dd055a.
Panic in the block layer if we see page-level corruption failures even
if it's not in DIAGNOSTIC mode, there's no reason to keep running.

Don't panic in the btree layer if we're doing a verify and there's a
decompression failure, verify should continue just like salvage.

Rename the SALVAGE_QUIET_ERR flag to SALVAGE_CORRUPT_OK, for clarity.