Allow some internal pages to be evicted without eviction getting aggressive.
Don't require eviction to visit pages twice before they can be evicted: in a large cache, it may take a very long time for eviction to move all the way through the cache, and this was keeping pages pinned for much longer than they needed to be.

Remember if some files are not generating many eviction candidates, and visit them less often than files with many evictable pages.