Merge split-page locking with compaction locking, and hold the lock slightly longer, through reconciliation wrapup.  (Holding the lock slightly longer isn't a fix for anything, all we care about is not seeing splits when we review the internal page's children, but it's simpler to hold it longer and release it at the same time as the compaction-required lock.)
Return immediately if we're trying to evict a page and it isn't possible.
This is the functionality of the old WT_SKIP_UPDATE_QUIT flag, which I
shouldn't have removed, but I think I can get the same effect by
checking that WT_EVICTION_LOCKED is set (and WT_SKIP_UPDATE_RESTORE is
not set), that's the "evicting but can't save-and-restore" combination.
I don't think this fixes any problems, but it means we'll quit sooner,
rather than later in some (rare) cases.

Merge __rec_txn_skip_chk() into __rec_child_deleted(), it's the only
caller now that __rec_txn_read() has been rewritten for new-split.