Rename "upd_skipped" to be "leave_dirty", it's the flag that sets if we can mark the page clean or not (allowing the page to be discarded).  Set the "leave_dirty" flag if there's an update we can't throw away or if there's an update that's unresolved.
Fix a bug where suffix compression could fail because the largest key
for a split block wasn't on the page, rather it was in a WT_UPDATE list
being saved/restored to a new block.

Fix a bug where a WT_UPDATE list being saved/restored to a new block was
associated with the wrong block because the key it came from was the
first key in a new block, after a split.  (The code to fix this bug in
column-store objects hasn't yet been written, so this commit only works
for row-store objects.)