SERVER-12615 Fix size calculation in JSReducer::_reduce
The old calculation didn't consider the size of the numeric-string fieldnames
in arrays. This error accumulated with each value added to the array. While
the new calculation doesn't consider the size of the current fieldname, it is
ok because it prevents the error from accumulating and we have plenty of room
between the User and Internal max sizes.

(cherry picked from commit 3e0aa5769dd58c727726c3397f9678d7f767b7ff)