Another run at handling removed overflow objects.  Michael suggested we could append a copy of the removed overflow value to the WT_UPDATE list we're going to save/restore as part of the eviction.  It's definitely a simpler approach and keeps all the magic inside of reconciliation.
Add code to __rec_txn_read() to append a WT_UPDATE entry with a copy of the
original overflow value and a transaction ID of WT_TXN_NONE to the update
list being saved.   Rework __rec_txn_read() to obviously quit sooner and
avoid all the extra work, not sure if the compiler needs the hints, but it
can't hurt.

Remove the WT_PAGE_OVERFLOW_RM page flag.
Remove the code to detect removed-overflow objects when instantiating a
disk image.
Remove __wt_multi_inmem_ovfl_rm() entirely, plus the pass over the disk
image to clean up removed overflow objects as part of the eviction split.