Bug when deepening the tree in a split, didn't connect the newly created internal child pages to the parent.
Bug when deepening the tree in a split, wasn't fixing up the right set
of WT_REF structures.

WT_REF.home needs to be volatile, we can't cache it in a register
otherwise we might not see the updates when a page splits.

Generalize the function to initialize a root WT_REF, call it from the
3 places that do that (salvage, root-page reconciliation, tree-open).

Simplify the code to look up the WT_REF slot.