Minor cleanups:
Set the checkpoint structure's version number when initializing.

WT_CKPT_UPDATE is a local flag and shouldn't be set in the passed-in
list.

Change __wt_block_extlist_write to discard blocks from the allocated
list, don't do it explicitly for all calls.

Don't assume the extent list name is of a specific length, it's a
combination of the application-specified checkpoint name in some cases.