Clean up the fast-delete code (part of doing a review after merging the new-split branch into develop).
Specific changes:

The code to rollback a fast-delete change has to loop, it's possible for
the instantiation to fail, returning the page to "deleted" status.  Add
a comment clarifying why we can't read a copy of the page (we don't
always have an open cursor on the tree when rolling back a transaction).

Add a new function __wt_page_parent_modify_set that dirties a parent
page.  It's only used in a couple of places, but it has a comment
explaining why dirtying a parent page is safe even in the face of
on-going splits that might be deepening the tree.

Note, these changes aren't sufficient for fast-deletes in the new-split
branch: while it's safe to update the single, instantiated page in-place
once it's set to an in-memory state (there is obviously an unresolved
transaction, the page isn't going anywhere), the problem is if the page
were to split.  In that case, the changes could be spread across
multiple physical pages, and the transaction unroll won't find them all.