Checkpointer test getting closer.
This test is designed to transactionally update multiple tables with
identical values. While in parallel executing checkpoints and verifying
that all tables in a checkpoint have identical content.

Supports row, column and LSM table types. Can either select multiple
of the same table type, or a mixture of table types.