Make it explicit that LSM cursors cannot read from checkpoints, at least until we make that work properly.
The main problem is that we don't have a copy of the LSM tree structure for each checkpoint, and as new chunks are added to the tree, ordinary cursors update and see more recent changes.  A secondary problem is ensuring that checkpoints correspond exactly to LSM chunk boundaries.  That way, a given chunk is either in or out of a checkpoint.