Changes to speed up retrieving a key from a leaf page.
If we roll-backward through a set of prefix compressed keys where the
prefix doesn't decrease, we don't need to roll forward through them, we
can start our roll-forward at the key that was the first we found with
the smallest prefix.

Don't bother allocating a temporary WT_ITEM if we don't need to, if we
are just taking the cell's bytes from the page, there's no additional
work that's necessary.