Rename WT_EVICTION_LOCKED to WT_EVICTING, we're going to set it when closing files as well as when evicting pages via the eviction server. The meaning is when that flag is set and there's an update we can't write, we quit reconciliation early because the page cannot be evicted.
Change the __evict_file() support for close to set the WT_EVICTING flag,
but not set the WT_SKIP_UPDATE_ERR flag.  This change means the EBUSY
we get when attempting to close a file that has updates that aren't yet
globally visible surfaces in the connection data handle code, rather
than resulting in a panic inside of reconciliation.

Move the reconciliation "panic if not allowed to skip updates" code back
into the code that first encounters those updates, rather than checking
at reconciliation wrapup -- it shouldn't be a change to the semantics,
but I think it's a little safer/cleaner/consistent.