Experimental change to avoid unpacking cells in order to access row-store leaf page keys.
If a page has no Huffman or prefix compression, and no overflow keys,
encode the key's page offset and size in the WT_ROW pointer and set a
flag on the page.  From that point on, when the flag is set, we can
go directly to the key instead of unpacking the key's cell each time.