Simplify the binary search of row-store internal pages by removing the test/branch for the 0th key of the page that we hit every time through the loop and instead detect that case after the loop exits; we may do an additional comparison in the case of searching down the left-side of the tree, but that's less likely and still probably worth doing.
Our collators are required to return -1, 0, or 1, tell the compiler so
it can use tests against a 0/1 register instead of the more complicated
negative/positive tests.

Separate out the common code for when an application collator is/is-not
configured.

Replace calls to WT_LEX_CMP_SKIP with WT_LEX_CMP where we know the
collator is configured, the compiler probably figured it all out, but
there's no reason to make it do that work.

Minor variable cleanup, there's no reason to set local variables until
they're actually useful; again, the compiler probably figured it out,
but no reason to make it do that work.