Add in a work around for the transaction visibility check in LSM.
The code uses the greater of the last update txn id or the chunk
max id for the visibility check.

This fixes my test case, but I suspect it's avoiding the issue, rather
than fixing it. (i.e hopefully we don't need to apply as many updates
to chunks, rather than having to wait longer to checkpoint).