Rename __cursor_search_clear __cursor_pos_clear; it's the code that clears a cursor's location, and is no longer unique to the search functions.
Change __cursor_error_resolve cursor error resolution to clear the
cursor's location even if we see a failure from the cursor "leave"
function (probably a bad return from releasing a page), keeping the
cursor's location isn't likely to make anything better.

sename __cursor_error_resolve to __cursor_reset, it's a combination of
"leave the API" and "clear the cursor's location"; now that it clears
the cursor location regardless of the return from the cursor "leave"
function, we can call it from the WT_CURSOR.reset code, they're the
same.