Do tree-append tracking in the cursor instead of in the btree handle.
Add a test for an append whenever we're inserting into an existing
insert list, try to catch inserts of lots of data at any point in
the tree.

Simplify when we decide a cursor might be appending to the tree, I
can't think of any approach that doesn't add significant complexity
to the row-store search function.

Reference #1124.