Add JSON loading to cursors and wt load utility. refs #740.
Allow set_key and set_value on JSON cursors.  The argument is a string
that is interpreted as JSON and must match the format created by a JSON
dump cursor or the 'wt dump -j' command.

Add -j option to 'wt load' utility to allow loading JSON.

Small adjustment to Python interface so that JSON cursor key/values inputs
are always treated as strings.

Added doc and Python tests for JSON cursors and wt load -j.