A couple of simple cleanups:
Allocate session handles directly into the structure memory we'll check
to decide if they need to be closed, that way if the called function
messes up, we still clean up.

Use the allocated session handle instead of the default session handle
when creating on the internal servers, that way we'll get better error
messages if something goes wrong.