Revision 01fe356 introduced an incompatibility, creating a database with 2.3.1, then trying to open it with develop without "create" gives the error "RUNDIR/WiredTiger.lock: No such file or directory". The problem is we were only creating the lock file if a "create" configuration was specified.
Rework database locking to always create the lock file if it doesn't
exist, and use the existence of the WiredTiger file to set the
create/no-create state of the database, instead of the lock file.

Don't acquire any file locks until we know there are no other threads
in the process with their own locks, closing a file handle discards all
locks held by the process.