Make fallocate/posix_fallocate more configurable.
When opening a file handle, add file-handle variables to note if
fallocate/posix_fallocate calls are available and if they require
locking.

Also, fallocate/posix_fallocate are known to fail at run-time based on
the underlying filesystem type. In that case, always return ENOTSUP from
fallocate/posix_fallocate, resetting the file-handle information and
falling back to ftruncate in the caller.

Remove the implementation of __wt_fallocate() using ftruncate, we may
have to re-acquire a lock when switching from a failed fallocate() call
to a ftruncate() call, when extending the file from the block manager,
so the caller has to do that work itself.

Reference #1267.