Add error-handling code to the insert split, if the split into the parent fails we can put things back and not panic.
We fixed a bug in 230cb5f to allocate a fresh copy of the key in the
first insert split WT_REF, change the error label so it's free'd.

Rename "current_ins" -> "ins", and "ins" -> "moved_ins", the special
one should have the special name, the one used in random loops should
have the generic name.

Callout the side-effect return, don't do it in the same paragraph as
local variable initialization.

Change warning/error messages to not include WT_PANIC in the errors we
ignore.