We documented that huffman_key and huffman_value took "none" as an argument, but they didn't.
Worse, if key was set but not value, we'd set value anyway, free of
charge, and if value was set but not key, we'd fail. I doubt this
is a problem (it's pretty surprising if anyone would set key but not
also set value).

Use __wt_config_gets_none() to support the "none" setting, re-work
the logic to ignore keys or values that are zero-length.

Reference #1417.