Don't try to write leaves from the sweep server.  Previously, this was done without locking the handle, and so could race with LSM discarding a handle.
We know the handle has been idle, so there is a good chance that a checkpoint has run since the last update and the write leaves was wasted effort.  If not, this change will keep the handle locked for longer preventing new opens, but it has been idle for a while anyway.

refs #1495, #1497 (maybe)