In recovery, track the maximum file ID in the metadata, regardless of whether there are any updates to roll forward.
Previously, we tracked the largest file ID that was updated in the logs being rolled forward.  It was usually the case that the most recently created file was also the most recently updated, so that calculation usually worked and wasn't detected until the repro in SERVER-17142 that created tables, did a clean shutdown and restart, then created more tables and did a dirty shutdown and restart, which was rolling forward updates into the wrong tables.

refs SERVER-17142, SERVER-17131(?)