There's a problem that went in in #1282, the key difference is that we are setting a split boundary at the end of the first page when there is more than a page worth of data. See also #1630 and #1631. This is an alternate approach to #1631:
The real change is to fallthrough into the split case if the next item
won't fit, callers of the split code can't handle failure from split,
it has to create enough room for the next item to be entered into the
buffer.

Note that now that SPLIT_MAX is back, we have to expect to see if when
we're finishing up a page's reconciliation, it just means everything
fit and we didn't have to split.

One additional fix: in SPLIT_TRACKING_OFF, we were growing the buffer
if we hadn't consumed half a page, it should be have a split size chunk.

One additional change: do the same calculation (if we haven't consumed
at least half a split-size chunk) when reaching the first boundary, and
in that case, grow the buffer.