Add a test case for checkpoint consistency.
We recently fixed a bug where updating a page after the write leaves
phase of a checkpoint had completed, then either checkpointed again or
closed the database. We would end up with inconsistent data. That was the
case because we weren't marking trees dirty all the time.

This test case reproduces the scenario so we don't reintroduce the failure.

Refs #1735 SUPPORT-1248, SERVER-17506